+++ lib/v6.dart
@@ -3,7 +3,7 @@
 import 'dart:typed_data';
 import 'data.dart';
 import 'parsing.dart';
-import 'uuid_util.dart';
+import 'rng.dart';
 
 class UuidV6 {
   final GlobalOptions? goptions;
@@ -14,8 +14,7 @@ class UuidV6 {
   /// Primarily sets up the seedBytes then generates the node id and clockseq
   void _init() {
     if (V6State.initialized) return;
-    Uint8List seedBytes = Function.apply(goptions?.rng ?? UuidUtil.mathRNG,
-        goptions?.positionalArgs ?? [], goptions?.namedArgs ?? {});
+    Uint8List seedBytes = goptions?.rng?.generate() ?? MathRNG().generate();
 
     // Per 4.5, create a 48-bit node id (47 random bits + multicast bit = 1)
     List<int> nodeId = [