+++ lib/v7.dart
@@ -2,7 +2,7 @@ import 'dart:typed_data';
 import 'package:uuid/data.dart';
 
 import 'parsing.dart';
-import 'uuid_util.dart';
+import 'rng.dart';
 
 class UuidV7 {
   final GlobalOptions? goptions;
@@ -50,8 +50,7 @@ class UuidV7 {
 
   /// _randomData() Generates a random data for v7 UUID random section
   List<int> _randomData() {
-    Uint8List seedBytes = Function.apply(goptions?.rng ?? UuidUtil.mathRNG,
-        goptions?.positionalArgs ?? [], goptions?.namedArgs ?? {});
+    Uint8List seedBytes = goptions?.rng?.generate() ?? MathRNG().generate();
 
     List<int> randomData = [
       seedBytes[0],