+++ test/uuid_test.dart
@@ -5,7 +5,7 @@ import 'dart:typed_data';
 import 'package:test/test.dart';
 import 'package:uuid/data.dart';
 import 'package:uuid/uuid.dart';
-import 'package:uuid/uuid_util.dart';
+import 'package:uuid/rng.dart';
 
 void main() {
   var uuid = const Uuid();
@@ -118,8 +118,7 @@ void main() {
   group('[Version 4 Tests]', () {
     test('Check if V4 is consistent using a static seed', () {
       var u0 = uuid.v4(options: {
-        'rng': UuidUtil.mathRNG,
-        'namedArgs': Map.fromIterables([const Symbol('seed')], [1])
+        'rng': MathRNG(seed: 1),
       });
       var u1 = 'a473ff7b-b3cd-4899-a04d-ea0fbd30a72e';
       expect(u0, equals(u1));
@@ -128,8 +127,7 @@ void main() {
     test('Consistency check with buffer', () {
       var buffer = Uint8List(16);
       uuid.v4buffer(buffer, options: {
-        'rng': UuidUtil.mathRNG,
-        'namedArgs': Map.fromIterables([const Symbol('seed')], [1])
+        'rng': MathRNG(seed: 1),
       });
 
       var u1 = 'a473ff7b-b3cd-4899-a04d-ea0fbd30a72e';
@@ -138,12 +136,10 @@ void main() {
 
     test('Using Objects', () {
       var regular = uuid.v4(options: {
-        'rng': UuidUtil.mathRNG,
-        'namedArgs': Map.fromIterables([const Symbol('seed')], [1])
+        'rng': MathRNG(seed: 1),
       });
       var obj = uuid.v4obj(options: {
-        'rng': UuidUtil.mathRNG,
-        'namedArgs': Map.fromIterables([const Symbol('seed')], [1])
+        'rng': MathRNG(seed: 1),
       });
 
       expect(obj.uuid, equals(regular));
@@ -363,7 +359,7 @@ void main() {
     });
 
     test('Explicit options produce expected id', () {
-      final rand = UuidUtil.mathRNG(seed: 1);
+      final rand = MathRNG(seed: 1).generate();
       var options = V7Options(1321651533573, rand);
       var id = uuid.v7(config: options);
 
@@ -399,7 +395,7 @@ void main() {
 
     test('Using buffers', () {
       var buffer = Uint8List(16);
-      final rand = UuidUtil.mathRNG(seed: 1);
+      final rand = MathRNG(seed: 1).generate();
       var options = V7Options(testTime, rand);
 
       var wihoutBuffer = uuid.v7(config: options);
@@ -409,7 +405,7 @@ void main() {
     });
 
     test('Using Objects', () {
-      final rand = UuidUtil.mathRNG(seed: 1);
+      final rand = MathRNG(seed: 1).generate();
       var options = V7Options(testTime, rand);
 
       var regular = uuid.v7(config: options);
@@ -569,7 +565,7 @@ void main() {
         ],
       }.entries) {
         test(testCase.key, () {
-          final rand = UuidUtil.mathRNG(seed: 1);
+          final rand = MathRNG(seed: 1).generate();
           final uuid =
               Uuid().v8(config: V8Options(testCase.value[0] as DateTime, rand));
           expect(uuid.toUpperCase(), equals(testCase.value[1]));